requirements.txt
anyio==4.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.6.15
cffi==1.17.1
chess==1.11.2
cssselect2==0.8.0
defusedxml==0.7.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pillow==11.2.1
pycparser==2.22
python-chess==1.999
python-telegram-bot==22.1
sniffio==1.3.1
tinycss2==1.4.0
webencodings==0.5.1
